+++ packages/frontend/src/components/ChatView/styles.module.scss
@@ -1,0 +1,101 @@
+@use '../../../scss/main_screen/_navbar';
+
+.chatAndNavbar {
+  display: flex;
+  flex-direction: column;
+
+  .chatNavbar {
+    @include navbar.navbar;
+    display: flex;
+    align-items: center;
+    padding: 0px;
+
+    .chatNavbarHeadingWrapper {
+      // Otherwise it will refuse to truncate the chat name
+      // and will make the parent overflow instead.
+      width: 0;
+      flex-grow: 1;
+    }
+    :global(.navbar-heading) {
+      position: relative;
+
+      width: fit-content;
+      // See `.chatNavbarHeadingWrapper`.
+      max-width: 100%;
+
+      display: flex;
+      align-items: center;
+      padding-inline-start: 5px;
+      margin-inline-end: 0px;
+      overflow-x: hidden;
+      white-space: nowrap;
+      text-overflow: ellipsis;
+      font-size: 18px;
+      font-weight: bold;
+      overflow: hidden;
+
+      h1,
+      h2,
+      h3,
+      h4,
+      h5,
+      h6 {
+        font-size: inherit;
+        font-weight: inherit;
+        display: block;
+        margin: 0;
+      }
+    }
+    :global(.navbar-heading-chat-info-button) {
+      // Fill the parent, overlay the siblings.
+      position: absolute;
+      inset: 0;
+      background-color: transparent;
+      border: none;
+    }
+
+    :global(.views) {
+      display: flex;
+      align-items: center;
+      min-height: 26px;
+    }
+
+    :global(.navbar-button) {
+      padding: 0;
+      width: 20px;
+      height: 20px;
+      border-radius: 0;
+      margin-inline-end: 12px;
+      &:hover {
+        background-color: var(--navBarButtonHover);
+      }
+    }
+
+    :global(#three-dot-menu-button) {
+      height: 20px;
+    }
+  }
+}
+
+.webxdcIcons {
+  line-height: 0;
+  margin-right: 12px;
+  img.webxdcIcon {
+    width: 25px;
+    height: 25px;
+    cursor: pointer;
+    border-radius: 3px;
+  }
+}
+
+.webxdcIconButton {
+  display: inline-block !important;
+  margin: 0;
+  margin-inline-end: 10px;
+  padding: 0 !important;
+  border-radius: 0 !important;
+  line-height: 0;
+  &:last-child {
+    margin-inline-end: 0;
+  }
+}
